Extract common rearrangeable list design
