+++ osu.Game/Overlays/Music/Playlist.cs
@@ -12,16 +12,11 @@ using osuTK;
 
 namespace osu.Game.Overlays.Music
 {
-    public class Playlist : RearrangeableListContainer<BeatmapSetInfo>
+    public class Playlist : OsuRearrangeableListContainer<BeatmapSetInfo>
     {
         public Action<BeatmapSetInfo> RequestSelection;
 
         public readonly Bindable<BeatmapSetInfo> SelectedSet = new Bindable<BeatmapSetInfo>();
-
-        /// <summary>
-        /// Whether any item is currently being dragged. Used to hide other items' drag handles.
-        /// </summary>
-        private readonly BindableBool playlistDragActive = new BindableBool();
 
         public new MarginPadding Padding
         {
@@ -33,14 +28,11 @@ namespace osu.Game.Overlays.Music
 
         public BeatmapSetInfo FirstVisibleSet => Items.FirstOrDefault(i => ((PlaylistItem)ItemMap[i]).MatchingFilter);
 
-        protected override RearrangeableListItem<BeatmapSetInfo> CreateDrawable(BeatmapSetInfo item) => new PlaylistItem(item)
+        protected override OsuRearrangeableListItem<BeatmapSetInfo> CreateOsuDrawable(BeatmapSetInfo item) => new PlaylistItem(item)
         {
             SelectedSet = { BindTarget = SelectedSet },
-            PlaylistDragActive = { BindTarget = playlistDragActive },
             RequestSelection = set => RequestSelection?.Invoke(set)
         };
-
-        protected override ScrollContainer<Drawable> CreateScrollContainer() => new OsuScrollContainer();
 
         protected override FillFlowContainer<RearrangeableListItem<BeatmapSetInfo>> CreateListFillFlowContainer() => new SearchContainer<RearrangeableListItem<BeatmapSetInfo>>
         {
